IA539 refactor: don't pass dispatch as param
- take redux store creation outside of index.js to /redux/store.js
- export dispatch
- import dispatch to requests.js i.o passing as param to API handlers
- put locationLimitMax in own file to avoid circular import
