+++ hat/assets/js/apps/Iaso/domains/orgUnits/constants/orgUnitConstants.js
@@ -1,0 +1,2 @@
+const locationLimitMax = 3000;
+export { locationLimitMax };
